+++ src/naive_solver.hpp
@@ -34,7 +34,7 @@ private:
 
         // TODO try remove first and contract second
         ContractionBackup backup = g_.contract_edge(u, v);
-        int loops = backup.u.inner_neighbors.count(v) - 1;
+        ll loops = backup.u.inner_neighbors.count(v) - 1;
         FlowPoly result = get_flow_poly_no_loops() * power(3, loops);
         g_.uncontract_edge(u, v, backup);
